rename ifconf to mode
