+++ engines/ifmode/ifconf_parser.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org>.
  */
 
-package ifconf
+package ifmode
 
 import "offscan/internal/argparser"
